fix: separate geo update e2e tests
